requirements.txt
streamlit
SpeechRecognition
pydub
google-generativeai
pdfplumber
python-dotenv
plotly
pandas
fpdf2
streamlit-echarts
numpy
requests
beautifulsoup4
streamlit-mic-recorder
